requirements.txt
# Python packages and dependencies

## Pip Packages
pillow

## Git Packages

